initial implementation of separation of Search Window
