+++ UI/Windows/FindReplaceWindow.xaml
@@ -1,0 +1,76 @@
+﻿<controls:MetroWindow
+             xmlns:controls="http://metro.mahapps.com/winfx/xaml/controls"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             x:Class="SPCode.UI.Windows.FindReplaceWindow"
+             WindowStartupLocation="CenterOwner" GlowBrush="{DynamicResource AccentColorBrush}" Background="{DynamicResource WhiteBrush}" ResizeMode="NoResize" ShowInTaskbar="False" Height="257" Width="496" HorizontalAlignment="Center" VerticalAlignment="Center" UseNoneWindowStyle="False" Topmost="True" TitlebarHeight="20">
+  <controls:MetroWindow.Resources>
+    <ResourceDictionary>
+      <ResourceDictionary.MergedDictionaries>
+        <ResourceDictionary Source="pack://application:,,,/MahApps.Metro;component/Styles/Controls.xaml" />
+        <ResourceDictionary Source="pack://application:,,,/MahApps.Metro;component/Styles/Fonts.xaml" />
+        <ResourceDictionary Source="pack://application:,,,/MahApps.Metro;component/Styles/Colors.xaml" />
+        <ResourceDictionary Source="pack://application:,,,/MahApps.Metro;component/Styles/Accents/BaseDark.xaml" />
+        <ResourceDictionary Source="pack://application:,,,/MahApps.Metro;component/Styles/Accents/Red.xaml" />
+        <ResourceDictionary Source="pack://application:,,,/MahApps.Metro;component/Styles/FlatButton.xaml" />
+        <ResourceDictionary>
+          <Storyboard x:Key="BlendOverEffect" Duration="00:00:00.5" Storyboard.TargetName="BlendEffectPlane" Storyboard.TargetProperty="Opacity">
+            <DoubleAnimationUsingKeyFrames>
+              <SplineDoubleKeyFrame Value="0.0" KeyTime="00:00:00.00" />
+              <SplineDoubleKeyFrame Value="0.5" KeyTime="00:00:00.25" />
+              <SplineDoubleKeyFrame Value="0.0" KeyTime="00:00:00.50" />
+            </DoubleAnimationUsingKeyFrames>
+          </Storyboard>
+          <Storyboard x:Key="FadeFindReplaceGridIn" Duration="00:00:00.2" Storyboard.TargetName="FindReplaceGrid">
+            <DoubleAnimation To="1" Duration="00:00:00.2" Storyboard.TargetProperty="Opacity" />
+          </Storyboard>
+          <Storyboard x:Key="FadeFindReplaceGridOut" Duration="00:00:00.2" Storyboard.TargetName="FindReplaceGrid">
+            <DoubleAnimation To="0" Duration="00:00:00.2" Storyboard.TargetProperty="Opacity" />
+          </Storyboard>
+          <Storyboard x:Key="EnableServerAnim">
+            <DoubleAnimation Duration="00:00:00.5" To="5" Storyboard.TargetName="ServerProgressBar" Storyboard.TargetProperty="Height" />
+            <BooleanAnimationUsingKeyFrames Storyboard.TargetName="ServerProgressBar" Storyboard.TargetProperty="IsIndeterminate">
+              <BooleanAnimationUsingKeyFrames.KeyFrames>
+                <BooleanKeyFrameCollection>
+                  <DiscreteBooleanKeyFrame KeyTime="00:00:00.25" Value="True" />
+                </BooleanKeyFrameCollection>
+              </BooleanAnimationUsingKeyFrames.KeyFrames>
+            </BooleanAnimationUsingKeyFrames>
+          </Storyboard>
+          <Storyboard x:Key="DisableServerAnim">
+            <DoubleAnimation Duration="00:00:00.5" To="0" Storyboard.TargetName="ServerProgressBar" Storyboard.TargetProperty="Height" />
+            <BooleanAnimationUsingKeyFrames Storyboard.TargetName="ServerProgressBar" Storyboard.TargetProperty="IsIndeterminate">
+              <BooleanAnimationUsingKeyFrames.KeyFrames>
+                <BooleanKeyFrameCollection>
+                  <DiscreteBooleanKeyFrame KeyTime="00:00:00.25" Value="False" />
+                </BooleanKeyFrameCollection>
+              </BooleanAnimationUsingKeyFrames.KeyFrames>
+            </BooleanAnimationUsingKeyFrames>
+          </Storyboard>
+        </ResourceDictionary>
+      </ResourceDictionary.MergedDictionaries>
+    </ResourceDictionary>
+  </controls:MetroWindow.Resources>
+  <Grid Name="FindReplaceGrid" Margin="0,0,0,-6">
+    <Grid.ColumnDefinitions>
+      <ColumnDefinition Width="10*"/>
+      <ColumnDefinition Width="21*"/>
+    </Grid.ColumnDefinitions>
+    <TextBox x:Name="FindBox" VerticalAlignment="Top" Margin="7,19,3,0" Height="45" controls:TextBoxHelper.Watermark="Find" VerticalScrollBarVisibility="Visible" TextChanged="SearchBoxTextChanged" KeyDown="SearchBoxKeyUp" Grid.ColumnSpan="2"/>
+    <TextBox x:Name="ReplaceBox" VerticalAlignment="Top" Margin="7,69,3,0" Height="45" controls:TextBoxHelper.Watermark="Replace" VerticalScrollBarVisibility="Visible" KeyDown="ReplaceBoxKeyUp" Grid.ColumnSpan="2"/>
+    <RadioButton x:Name="NSearch_RButton" HorizontalAlignment="Left" VerticalAlignment="Top" Margin="7,119,0,0" Content="Normal Search" IsChecked="True" Foreground="{DynamicResource BlackBrush}" Click="SearchButtonClicked" />
+    <RadioButton x:Name="WSearch_RButton" HorizontalAlignment="Left" VerticalAlignment="Top" Margin="7,139,0,0" Content="Match Whole Words" Foreground="{DynamicResource BlackBrush}" />
+    <RadioButton x:Name="ASearch_RButton" HorizontalAlignment="Left" VerticalAlignment="Top" Margin="7,159,0,0" Content="Advanced Search (\r, \n, \t)" Foreground="{DynamicResource BlackBrush}" Grid.ColumnSpan="2" />
+    <RadioButton x:Name="RSearch_RButton" HorizontalAlignment="Left" VerticalAlignment="Top" Margin="7,178,0,0" Content="Regex Search" Foreground="{DynamicResource BlackBrush}" />
+    <ComboBox x:Name="FindDestinies" HorizontalAlignment="Left" VerticalAlignment="Top" Margin="22,119,0,0" Width="150" Grid.Column="1">
+      <ComboBoxItem x:Name="MenuFR_CurrDoc" IsSelected="True" Content="Current Document" />
+      <ComboBoxItem x:Name="MenuFR_AllDoc" Content="All Open Documents" />
+    </ComboBox>
+    <Button x:Name="Find_Button" HorizontalAlignment="Right" VerticalAlignment="Top" Margin="0,119,3,0" Content="Find (F3)" Width="150" Click="SearchButtonClicked" Style="{DynamicResource MetroFlatButton}" Grid.Column="1"/>
+    <controls:SplitButton x:Name="ReplaceButton" HorizontalAlignment="Right" VerticalAlignment="Top" Margin="0,151,3,0" Width="150" Click="ReplaceButtonClicked" Grid.Column="1"/>
+    <Button x:Name="Count_Button" HorizontalAlignment="Right" VerticalAlignment="Top" Margin="0,180,3,0" Content="Count" Width="150" Click="CountButtonClicked" Style="{DynamicResource MetroFlatButton}" Grid.Column="1"/>
+    <CheckBox x:Name="CCBox" HorizontalAlignment="Left" VerticalAlignment="Top" Margin="22,149,0,0" Content="Case Sensitive" Foreground="{DynamicResource BlackBrush}" Grid.Column="1" />
+    <CheckBox x:Name="MLRBox" HorizontalAlignment="Left" VerticalAlignment="Top" Margin="22,174,0,0" Content="Multiline Regex" Foreground="{DynamicResource BlackBrush}" Grid.Column="1" />
+    <TextBlock x:Name="FindResultBlock" HorizontalAlignment="Left" VerticalAlignment="Bottom" Margin="20,0,0,24" Text="" Width="Auto" FontSize="14" Foreground="{DynamicResource WhiteBrush}" IsHitTestVisible="False" />
+  </Grid>
+</controls:MetroWindow>
